requirements.txt
altgraph==0.17
argcomplete==1.10.0
asgiref==3.2.3
asn1crypto==1.2.0
astroid==2.3.3
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
colorama==0.4.3
conda==4.7.12
conda-package-handling==1.6.0
crypto==1.4.1
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
Django==3.0.3
docx2txt==0.8
EbookLib==0.17.1
et-xmlfile==1.0.1
extract-msg==0.23.1
future==0.18.2
idna==2.8
IMAPClient==2.1.0
isort==4.3.21
jdcal==1.4.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.0
matplotlib==3.1.2
mccabe==0.6.1
menuinst==1.4.16
MouseInfo==0.1.2
Naked==0.1.31
netifaces==0.10.9
numpy==1.17.4
numpy-stl==2.10.1
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.3
pandas==1.0.1
pbkdf2==1.3
pdfkit==0.6.1
pdfminer.six==20181108
pefile==2019.4.18
Pillow==7.0.0
PyAutoGUI==0.9.48
pycosat==0.6.3
pycparser==2.19
pycryptodome==3.9.4
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.4.4
PyMsgBox==1.0.7
pynput==1.6.7
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyPdf==1.13
PyPDF2==1.26.0
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-pptx==0.6.18
python-utils==2.3.0
PyTweening==1.0.3
pytz==2019.3
pywin32==223
pywin32-ctypes==0.2.0
PyYAML==5.3
requests==2.22.0
ruamel-yaml==0.15.46
selenium==3.141.0
shellescape==3.8.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
textract==1.6.3
tqdm==4.36.1
typed-ast==1.4.1
tzlocal==1.5.1
urllib3==1.24.2
virtualenv==16.7.8
wifi==0.3.8
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.7
